moving some functions into a separate file
