+++ test/stickerbook.test.js
@@ -7,48 +7,7 @@ const sinon = require('sinon');
 const Stickerbook = require('../src/stickerbook');
 const historyFixture = require('./data/historyFixture.json');
 
-const images = {
-  star: 'data:image/gif;base64,R0lGODlhEAAQAMQAAORHHOVSKudfOulrSOp3WOyDZu6QdvCchPGolfO0o/XBs/fNwfjZ0frl3/zy7////wAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAACH5BAkAABAALAAAAAAQABAAAAVVICSOZGlCQAosJ6mu7fiyZeKqNKToQGDsM8hBADgUXoGAiqhSvp5QAnQKGIgUhwFUYLCVDFCrKUE1lBavAViFIDlTImbKC5Gm2hB0SlBCBMQiB0UjIQA7',
-  dot: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg==',
-  box: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAAECAYAAACk7+45AAAAFElEQVQYV2NkYGD4z8DAwMCImwEASjQEAY/F9H0AAAAASUVORK5CYII='
-};
-
-const createValidConfig = () => {
-  return {
-    container: document.createElement('div'),
-    stickers: [
-      images.star,
-      images.dot
-    ],
-    brush: {
-      widths: [1, 10],
-      enabled: ['eraser', 'bitmap', 'fill', 'marker', 'pencil', 'spray'],
-      colors: ['#0000FF', '#FF0000'],
-    },
-    background: {
-      enabled: [images.dot],
-      default: null
-    },
-    stickerControls: {
-      cornerColor: 'rgba(0, 0, 0, 0.5)',
-      cornerSize: 20
-    },
-    useDefaultEventHandlers: true,
-    mobileEnabled: true
-  };
-};
-
-const createStickerbook = () => {
-  var stickerbook = new Stickerbook(createValidConfig());
-
-  // fake the dimensions of each of the canvas elements
-  stickerbook.backgroundManager._canvas.width = 100;
-  stickerbook.backgroundManager._canvas.height = 100;
-  stickerbook._canvas.lowerCanvasEl.width = 100;
-  stickerbook._canvas.lowerCanvasEl.height = 100;
-
-  return stickerbook;
-};
+const { images, createValidConfig, createStickerbook } = require('./helpers.js');
 
 describe('Stickerbook', () => {
   it('does not present config as part of public API', () => {
